Colab Notebook and csv files
Uploaded 2 datasets (csv files) and a notebook with a simple model for a text binary classification problem. Both csv files have different information but have text content of tweets and the a binary tag (disaster or not, relevant or not). The .ipynb works with the tweets.csv file, but you have to upload it to your colab session for it to work. Small changes have to be adjusted for it to work with the other file (different named columns).